[WasmFS] Report errors from `getSize` (#17783)
Update the return type of `getSize` from `size_t` to `off_t`, which allows for
negative error codes. `off_t` is more correct anyway, since in 32-bit mode
it is possible for a file size to be greater than the max value of `size_t`.
Also update the argument to `getSize` from `size_t` to `off_t` to match.

I manually verified that this error handling works with an injected error, but I
was not able to create a test that could trigger the error naturally.

+++ system/lib/wasmfs/backends/node_backend.cpp
@@ -137,7 +137,7 @@ public:
     : DataFile(mode, backend), state(path) {}
 
 private:
-  size_t getSize() override {
+  off_t getSize() override {
     // TODO: This should really be using a 64-bit file size type.
     uint32_t size;
     if (state.isOpen()) {
@@ -151,10 +151,10 @@ private:
         return 0;
       }
     }
-    return size_t(size);
+    return off_t(size);
   }
 
-  int setSize(size_t size) override {
+  int setSize(off_t size) override {
     WASMFS_UNREACHABLE("TODO: implement NodeFile::setSize");
   }
